requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
pydantic==1.10.7
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.4
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
